backend/requirements.txt
# FastAPI e servidor
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# MongoDB
motor==3.3.2
pymongo==4.6.1

# Autenticação e segurança
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validação e configuração
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilitários
python-dotenv==1.0.0

